mz: migrate API requests into `api` module
Begin separating out the code that makes API requests from the "UI
layer" that handles interacting with the user via the terminal.
