+++ src/mz/src/lib.rs
@@ -81,3 +81,7 @@
 #![warn(clippy::from_over_into)]
 // END LINT CONFIG
 
+pub mod api;
+pub mod configuration;
+pub mod utils;
+pub mod vault;
